+++ src/transform.rs
@@ -21,6 +21,7 @@ use crate::{
 /// This is a strongly typed wrapper around a [`Matrix3`], where that matrix
 /// describes how to map between units.
 #[repr(transparent)]
+#[cfg_attr(feature = "facet", derive(facet_derive::Facet))]
 pub struct Transform2<Src: Unit, Dst: Unit> {
     /// Underlying matrix.
     pub matrix: Matrix3<Src::Scalar>,
@@ -32,6 +33,7 @@ pub struct Transform2<Src: Unit, Dst: Unit> {
 /// This is a strongly typed wrapper around a [`Matrix4`], where that matrix
 /// describes how to map between units.
 #[repr(transparent)]
+#[cfg_attr(feature = "facet", derive(facet_derive::Facet))]
 pub struct Transform3<Src: Unit, Dst: Unit> {
     /// Underlying matrix.
     pub matrix: Matrix4<Src::Scalar>,